Create honey_selection.py
monitors directory for any changes(additions or deletions) of any files within specified directory. Upon detection of new/altered files, this will run through adding the contents of the files to the database. This will also crawl the suspected honeypot URL to determine if it is a dead link or not. More improvements to this file are coming soon.